prog fixes

+++ programs/all-testers/M4/testers/funcs/return_by_val/float.go
@@ -2,7 +2,7 @@
 
 package main
 
-func somefunc(a float64) float {
+func somefunc(a float64) float64 {
 	return a
 }
